% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSetDocumentation.R
\docType{data}
\name{hair1958}
\alias{hair1958}
\title{hair1958}
\format{A data frame with 58 observations on 23 variables}
\source{
This is not the real source just a test:  \url{http://www.fpl.fs.fed.us/documnts/pdf2008/fpl_2008_skog001.pdf}
}
\usage{
hair1958
}
\description{
Data I am just using right now.
}
\keyword{datasets}

