% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplemathr.R
\docType{package}
\name{simplemathr-package}
\alias{simplemathr-package}
\title{The most incredible editable simplemathr package}
\description{
Miscelaneous functions for testing code in class.
}
\details{
Explain all kinds of stuff here.
}
\section{My Custom Section}{
 This is just me blathering...yo que se
}
\keyword{package}

