% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSetDocumentation.R
\docType{data}
\name{testWoodcarb}
\alias{testWoodcarb}
\title{testWoodcarb}
\format{A data frame with 151 observations on 75 variables}
\usage{
testWoodcarb
}
\description{
Data I am just using right now.
}
\keyword{datasets}

