% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Two-Sample Equal Variance t-Test}
\usage{
ttest(x, y, alpha = 1/20, alternative = "two.sided", ...)
}
\arguments{
\item{x}{a non-empty numeric vector.}

\item{y}{a non-empty numeric vector.}

\item{alpha}{value to test hypothesis.}

\item{alternative}{hypothesis (can take on one of three values).}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
the answers!
}
\description{
Two-Sample Equal Variance t-Test
}
\examples{
T1 <- ttest(x = rnorm(300.5, 1), y = rnorm(30, 1, 1), alternative = "less")
T1
print(T1)
plot(T1)
Mx1x2 <- matrix(rnorm(60, 0, 1), nrow = 30)
class(Mx1x2)
TM <- ttest(x = Mx1x2)
TM
plot(TM)
ttest(x = Mx1x2) # Test that a matrix will work
mylist <- list(xl = rnorm(30), yl = rnorm(30))
class(mylist)
ttest(x = mylist, alternative = "greater")
plot(ttest(mylist))

}
